UVSC to git
