--- .plastic/plastic.selector
@@ -1,4 +1,0 @@
-repository "TodoRPG@anglee6915@cloud"
-  path "/"
-    br "/main"
-    co "/main"
