+++ Assets/Scenes/Sample.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 1c08b3733d6ded44ebc89d5598c4c5cd
-folderAsset: yes
+guid: 84513ef2bfa14ec4ebbaf8b052d7fc93
 DefaultImporter:
   externalObjects: {}
   userData:
