+++ Assets/Settings/Lit2DSceneTemplate.scenetemplate.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 1c08b3733d6ded44ebc89d5598c4c5cd
-folderAsset: yes
-DefaultImporter:
+guid: d03ed43fc9d8a4f2e9fa70c1c7916eb9
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
