+++ Assets/Settings/Renderer2D.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 1c08b3733d6ded44ebc89d5598c4c5cd
-folderAsset: yes
-DefaultImporter:
+guid: 424799608f7334c24bf367e4bbfa7f9a
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
