+++ Assets/Settings/Scenes/URP2DSceneTemplate.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 1c08b3733d6ded44ebc89d5598c4c5cd
-folderAsset: yes
+guid: 2cda990e2423bbf4892e6590ba056729
 DefaultImporter:
   externalObjects: {}
   userData:
