+++ Assets/Settings/UniversalRP.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 1c08b3733d6ded44ebc89d5598c4c5cd
-folderAsset: yes
-DefaultImporter:
+guid: 681886c5eb7344803b6206f758bf0b1c
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
