+++ Assets/UniversalRenderPipelineGlobalSettings.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ec09bf2a4fe6ff543a703caae8d963c4
+guid: 93b439a37f63240aca3dd4e01d978a9f
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
