+++ Assets/UniversalRenderPipelineGlobalSettings.asset.meta.private.0
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 1c08b3733d6ded44ebc89d5598c4c5cd
-folderAsset: yes
-DefaultImporter:
+guid: ec09bf2a4fe6ff543a703caae8d963c4
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
