+++ Assets/UniversalRenderPipelineGlobalSettings.asset.meta.private.0.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 1c08b3733d6ded44ebc89d5598c4c5cd
-folderAsset: yes
+guid: 70dd9d407562dcf4fbfd478242389d01
 DefaultImporter:
   externalObjects: {}
   userData:
