+++ Assets/UniversalRenderPipelineGlobalSettings.asset.private.0.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 1c08b3733d6ded44ebc89d5598c4c5cd
-folderAsset: yes
+guid: f31176067c9471644953ff79a3263b3a
 DefaultImporter:
   externalObjects: {}
   userData:
