+++ Assets/character/New 2D Renderer Data.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 1c08b3733d6ded44ebc89d5598c4c5cd
-folderAsset: yes
-DefaultImporter:
+guid: 64c75f09225e87141b9740d9cd7dd54a
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
