+++ ProjectSettings/PackageManagerSettings.asset
@@ -31,6 +31,6 @@ MonoBehaviour:
   m_RegistryInfoDraft:
     m_Modified: 0
     m_ErrorMessage: 
-    m_UserModificationsInstanceId: -888
-    m_OriginalInstanceId: -890
+    m_UserModificationsInstanceId: -838
+    m_OriginalInstanceId: -840
   m_LoadAssets: 0
