+++ ProjectSettings/ProjectSettings.asset.private.0
@@ -4,7 +4,7 @@
 PlayerSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 26
-  productGUID: 5c226897370c68047a04c97c5f916af1
+  productGUID: f735cbbd31907744785de7da9779079f
   AndroidProfiler: 0
   AndroidFilterTouchesWhenObscured: 0
   AndroidEnableSustainedPerformanceMode: 0
