+++ ignore.conf
@@ -24,8 +24,6 @@ logs
 ^*.private.[0-9]+.meta$
 .vs
 .vscode
-.git
-.gitignore
 .idea
 .gradle
 ExportedObj
